+++ libs/eventually-broker/src/channels/HttpPostPushChannel.ts
@@ -2,12 +2,17 @@ import { log } from "@rotorsoft/eventually";
 import axios from "axios";
 import { PushChannel } from "../interfaces";
 
+const TIMEOUT = 10000;
+
 export const HttpPostPushChannel = (endpoint: URL): PushChannel => {
   return {
+    label: "",
     init: () => undefined,
     push: async (event) => {
       try {
-        const response = await axios.post(endpoint.href, event);
+        const response = await axios.post(endpoint.href, event, {
+          timeout: TIMEOUT
+        });
         const { status, statusText } = response;
         return { status, statusText };
       } catch (error) {